update component folders through flattening and renamed button components for consistency
