+++ ngx-uswds-workspace/projects/ngx-uswds-lib/src/components/button/ngx-uswds-component-button-lib.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { UswdsButton } from './uswds-button';
+import { UswdsButton } from './ngx-uswds-component-button-lib';
 
 describe('UswdsButton', () => {
   let component: UswdsButton;
