+++ ngx-uswds-workspace/projects/ngx-uswds-lib/src/components/button/ngx-uswds-component-button-lib.ts
@@ -3,8 +3,8 @@ import { Component, Input } from '@angular/core';
 @Component({
   selector: 'lib-uswds-button',
   imports: [],
-  templateUrl: './uswds-button.html',
-  styleUrl: './uswds-button.css',
+  templateUrl: './ngx-uswds-component-button-lib.html',
+  styleUrl: './ngx-uswds-component-button-lib.css',
 })
 export class UswdsButton {
   @Input() text: string = "EMPTY";
