+++ ngx-uswds-workspace/projects/ngx-uswds-lib/src/components/card-footer/ngx-uswds-component-card-footer-lib.ts
