+++ ngx-uswds-workspace/projects/ngx-uswds-lib/src/components/card-header/ngx-uswds-component-card-header-lib.ts
@@ -1,5 +1,5 @@
 import { Component, Input} from '@angular/core';
-import { cardHeaderLevelOptions } from '../types/ngx-uswds-type-card-header';
+import { cardHeaderLevelOptions } from './ngx-uswds-type-card-header';
 
 @Component({
   selector: 'uswds-card-header',
