+++ ngx-uswds-workspace/projects/ngx-uswds-lib/src/components/card-media/ngx-uswds-component-card-media-lib.html
