+++ ngx-uswds-workspace/projects/ngx-uswds-lib/src/public-api.ts
@@ -2,13 +2,13 @@
  * Public API Surface of ngx-uswds-lib
  */
 
-export * from './components/uswds-button/uswds-button';
+export * from './components/button/ngx-uswds-component-button-lib';
 
 // Card Components
-export * from './components/card/component/ngx-uswds-component-card-lib';
-export * from './components/card-body/component/ngx-uswds-component-card-body-lib';
-export * from './components/card-footer/component/ngx-uswds-component-card-footer-lib';
-export * from './components/card-header/component/ngx-uswds-component-card-header-lib';
-export * from './components/card-media/component/ngx-uswds-component-card-media-lib';
+export * from './components/card/ngx-uswds-component-card-lib';
+export * from './components/card-body/ngx-uswds-component-card-body-lib';
+export * from './components/card-footer/ngx-uswds-component-card-footer-lib';
+export * from './components/card-header/ngx-uswds-component-card-header-lib';
+export * from './components/card-media/ngx-uswds-component-card-media-lib';
